// src/lib/rag/localRagEmbeddings.generated.ts
export type TableDetailEmbedding = {
  id: string;
  name: string;
  description: string;
  embedding: number[];
};

export type FewShotEmbedding = {
  id: string;
  query: string;
  answer: string;
  embedding: number[];
};